+++ src/app/my-page/page.tsx
@@ -1,33 +1,24 @@
-'use client';
-import { useRouter } from 'next/navigation';
+import 'server-only';
 
-import { GearFillIcon } from '~/components/Icon';
-import { TopNavigation } from '~/components/TopNavigation';
-import { createIdCard } from '~/mocks/idCard/idCard.mock';
+import { getCommunityMyIdCardDetail } from '~/api/domain/idCard.api';
 import { IdCard } from '~/modules/IdCard';
+import { IdCardEditButton } from '~/modules/IdCardEditButton';
+import { PlanetCreationButton } from '~/modules/PlanetCreationButton';
 
-const MyPage = () => {
-  const router = useRouter();
-  // TODO: 해당 행성에서 내 주민증 정보 요청 api 추가 예정
-  const { idCardId, nickname, aboutMe, characterType, keywords } = createIdCard();
+const MyPage = async () => {
+  // TODO: 행성전환 store에서 가져오기
+  const communityId = 123;
+  const { idCardDetailsDto } = await getCommunityMyIdCardDetail(communityId);
+  const { idCardId, nickname, aboutMe, characterType, keywords } = idCardDetailsDto;
   const keywordTitles = keywords.map(keyword => keyword.title);
 
-  const onClickGearFill = () => {
-    router.push('/my-page/config');
-  };
-
   return (
-    <main>
-      <TopNavigation bottomBorderColor="bg-grey-100">
-        <TopNavigation.Left>커뮤니티 이동 컴포넌트</TopNavigation.Left>
-        <TopNavigation.Title></TopNavigation.Title>
-        <TopNavigation.Right>
-          <GearFillIcon onClick={onClickGearFill} />
-        </TopNavigation.Right>
-      </TopNavigation>
-      <div className="mt-[86px]">
-        {/* TODO: mt[-50px]: 레이아웃 조정하면서 빼야하는 top-navigation height */}
-        <h2 className="mb-16pxr text-h3 text-grey-800">내 주민증</h2>
+    <main className="pt-35pxr">
+      <div className="mx-layout-l">
+        <div className="mb-16pxr flex w-full justify-between">
+          <h2 className="text-h3 text-grey-800">내 주민증</h2>
+          <IdCardEditButton />
+        </div>
         <IdCard
           idCardId={idCardId}
           nickname={nickname}
@@ -36,10 +27,8 @@ const MyPage = () => {
           keywordTitles={keywordTitles}
         />
       </div>
-      <div className="mt-28pxr">
-        {/* 해당 block는 좌표 관련 기획이 마무리되면 추가할 예정 */}
-        <div>행성 만들기</div>
-        <div>좌표로 행성 찾기</div>
+      <div className="mx-layout-sm mt-28pxr">
+        <PlanetCreationButton />
       </div>
     </main>
   );
